feat: M2 Batch B — delete Unit3D family + cascade files, update surviving refs
Delete ~10,000 LOC of legacy view-layer code now replaced by the
Simulation + View layer architecture:

Phase 3a - Update surviving files:
- CardFactory: strip spell/summon execution, keep query API only
- CardCatalog: use UnitDefinitions for separation_radius
- GhostUnit3D/SpawnPreview: use UnitDefinitions for flight altitude
- SpatialGrid/DesyncDetector: duck-typed property access via Get()
- CombatEvent: duck-typed Team access
- All 17 Targeting/ files: Unit3D → Node3D in method signatures

Phase 3b - Relocate SpawnRevealComponent to View/ namespace

Phase 3c - Delete Unit3D family:
- Unit3D (2,285 LOC), MeleeUnit3D, RangedUnit3D, DucklingUnit3D
- UnitHealth, UnitMovement, UnitDebugService

Phase 3d - Delete Unit3D dependencies:
- UnitSteering (462 LOC), UnitSpawner (419 LOC), UnitSummon, SummonResult

Phase 3e - Delete Cards/Effects + SpellCard + SpellBuilder:
- Entire Effects/ directory (14 files), SpellCard, SpellBuilder

Phase 3f - Delete Projectile3D (1,128 LOC)

Also includes M1 Tier 1 leftovers:
- DamageSystem, ModifierService, ProjectileService deletions
- StatModifier + TriggerCondition relocated to Stats/

+++ scripts/csharp/Multiplayer/Authority/HostEventBroadcaster.cs
@@ -32,7 +32,7 @@ public class HostEventBroadcaster : ISimEventVisitor
             _broadcaster.Broadcast(new UnitSpawned(
                 unit.NetworkId,
                 e.CatalogId,
-                unit.Team,
+                (int)unit.Team,
                 pos,
                 _state.FrameNumber,
                 null, null,
